feat(Countly.helpers): rename to Telemetry.helpers
